Add optional update checking on each launch of the application
This new feature uses a simple API powered by AWS Lambda that checks
the Github project for a newer release than the current one which is
passed as a query parameter. It returns some information about the
version so some of it can be displayed to user and the URL for that
release can be opened if wanted.